+++ NEWS.md
@@ -1,6 +1,7 @@
 # snakecase 0.8.3
 
 * replaced `parsing_option`s 3 and 4 with 5 and 6. 
+* removed __purrr__ dependecy
 
 # snakecase 0.8.2.9002